% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_scalar}
\alias{as_scalar}
\title{Tag an atomic vector with a single element as class 'scalar'.  When output
to JSON it will be output as a scalar not a vector}
\usage{
as_scalar(x)
}
\arguments{
\item{x}{atomic vector with length = 1}
}
\value{
If x is an atomic vector with length = 1, then object is returned
with class 'scalar', otherwise object is returned unmodified
}
\description{
Tag an atomic vector with a single element as class 'scalar'.  When output
to JSON it will be output as a scalar not a vector
}
