% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson.R
\name{from_geojson_opts}
\alias{from_geojson_opts}
\title{Options for reading in GeoJSON}
\usage{
from_geojson_opts(
  type = c("sf", "sfc"),
  property_promotion = c("string", "list"),
  property_promotion_lgl_as_int = c("integer", "string")
)
}
\arguments{
\item{type}{'sf' or 'sfc'}

\item{property_promotion}{What is the most general container type to use when
properties differ across a FEATURECOLLECTION?  E.g. if the property
exists both as a numeric and a string, should all values be promoted
to a 'string', or contained as different types in a 'list'.
Default: 'string' will behave like 'geojsonsf'}

\item{property_promotion_lgl_as_int}{when promoting properties into a string,
should logical values beccome strings e.g. "TRUE" or integers
e.g. "1".  Default: "integer" in order to match \code{geojsonsf} packages}
}
\value{
named list
}
\description{
Options for reading in GeoJSON
}
