% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{from_json_file}
\alias{from_json_file}
\title{Convert JSON to R}
\usage{
from_json_file(filename, opts = list(), ...)
}
\arguments{
\item{filename}{full path to text file containing JSON.}

\item{opts}{Named list of options for parsing. Usually created by \code{from_opts()}}

\item{...}{Other named options can be used to override any options in \code{opts}.
The valid named options are identical to arguments to \code{\link[=from_opts]{from_opts()}}}
}
\value{
R object
}
\description{
Convert JSON to R
}
\examples{
\dontrun{
from_json_file("myfile.json")
}

}
\seealso{
Other JSON Parsers: 
\code{\link{from_json_conn}()},
\code{\link{from_json_raw}()},
\code{\link{from_json_str}()},
\code{\link{from_ndjson_file_as_df}()}
}
\concept{JSON Parsers}
