% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-yyjson.R
\name{from_json_str}
\alias{from_json_str}
\title{Convert JSON to R}
\usage{
from_json_str(str, opts = from_opts(), ...)
}
\arguments{
\item{str}{a single character string}

\item{opts}{Named list of options for parsing. Usually created by \code{from_opts()}}

\item{...}{Other named options can be used to override any options in \code{opts}.
The valid named options are identical to arguments to \code{\link[=from_opts]{from_opts()}}}
}
\value{
R object
}
\description{
Convert JSON to R
}
\examples{
\dontrun{
from_json_str(str, opts = iris_opts(int64 = 'string'))
}

}
\seealso{
Other JSON Parsers: 
\code{\link{from_json_conn}()},
\code{\link{from_json_file}()},
\code{\link{from_ndjson_file_as_df}()}
}
\concept{JSON Parsers}
