% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-yyjson.R
\name{from_ndjson_file_as_df}
\alias{from_ndjson_file_as_df}
\alias{from_ndjson_file_as_list}
\title{Parse an ndjson file to a data.frame or list}
\usage{
from_ndjson_file_as_df(
  filename,
  nread = -1,
  nskip = 0,
  nprobe = 100,
  opts = list(),
  ...
)

from_ndjson_file_as_list(filename, nread = -1, nskip = 0, opts = list(), ...)
}
\arguments{
\item{filename}{text file or gzipped file}

\item{nread}{number of records to read. Default: -1 (reads all rows)}

\item{nskip}{number of records to skip before starting to read. Default: 0}

\item{nprobe}{Number of lines to read to determine types for data.frame
columns.  Default: 100. Only valid for \code{from_ndjson_file_as_df()}}

\item{opts}{Named list of options for parsing. Usually created by \code{from_opts()}}

\item{...}{Other named options can be used to override any options in \code{opts}.
The valid named options are identical to arguments to \code{\link[=from_opts]{from_opts()}}}
}
\value{
data.frame

list
}
\description{
If reading as data.frame, each row of ndjson becomes a row in the data.frame.
If reading as a list, then each row becomes an element in the list.
}
\details{
If parsing ndjson to a data.frame it is usually better if the json objects
are consistent from line-to-line.  Type inference for the data.frame is done
during initialisation by reading through \code{nprobe} lines.  Warning: if
there is a type-mismatch further into the file than it is probed, then you
will get missing values in the data.frame.

No flattening of the namespace is done i.e. nested object remain nested.
}
\examples{
\dontrun{
from_ndjson_file_as_df("flights.ndjson", nskip = 1000, nread = 200)
from_ndjson_file_as_list("flights.ndjson", nskip = 1000, nread = 200)
}

}
\seealso{
Other JSON Parsers: 
\code{\link{from_json_conn}()},
\code{\link{from_json_file}()},
\code{\link{from_json_raw}()},
\code{\link{from_json_str}()}

Other JSON Parsers: 
\code{\link{from_json_conn}()},
\code{\link{from_json_file}()},
\code{\link{from_json_raw}()},
\code{\link{from_json_str}()}
}
\concept{JSON Parsers}
