% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson.R
\name{opts_write_geojson}
\alias{opts_write_geojson}
\title{Options for writing from \code{sf} object to \code{GeoJSON}}
\usage{
opts_write_geojson()
}
\value{
Named list of options specific to writing GeoJSON
}
\description{
Currently no options available.
}
\examples{
# Create a set of options to use when writing geojson
opts_write_geojson()
}
