% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-ndjson.R
\name{read_dataset_ndjson_raw}
\alias{read_dataset_ndjson_raw}
\title{Parse a dataset-NDJSON file to a data.frame}
\usage{
read_dataset_ndjson_raw(rvec, nskip = 1, colspec = NULL, opts = list(), ...)
}
\arguments{
\item{rvec}{raw vector}

\item{nskip}{Number of records to skip before starting to read. Default: 0
(skip no data)}

\item{colspec}{data.frame of column spec. Must have a row for each data item
expected in the dataset.  Must include a 'name' and 'dataType' column}

\item{opts}{Named list of options for parsing. Usually created by \code{opts_read_json()}}

\item{...}{Other named options can be used to override any options in \code{opts}.
The valid named options are identical to arguments to \code{\link[=opts_read_json]{opts_read_json()}}}
}
\description{
Parse a dataset-NDJSON file to a data.frame
}
