% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson.R
\name{read_geojson_conn}
\alias{read_geojson_conn}
\title{Parse geoJSON from an R connection object.}
\usage{
read_geojson_conn(conn, opts = list(), ...)
}
\arguments{
\item{conn}{connection object.  e.g. \code{url('https://jsonplaceholder.typicode.com/todos/1')}}

\item{opts}{Named list of GeoJSON-specific options. Usually created
with \code{opts_read_geojson()}.
Default: empty \code{list()} to use the default options.}

\item{...}{Any extra named options override those in GeoJSON-specific options
- \code{opts}}
}
\value{
R object
}
\description{
Currently, this is not very efficient as the entire contents of the connection are
read into R as a string and then the JSON parsed from there.
}
\details{
For uncompress geojson files it is faster to use
\code{read_geojson_file()}.
}
\examples{
if (interactive()) {
  read_geojson_conn(gzfile("example.geojson.gz"))
}


}
\seealso{
Other JSON Parsers: 
\code{\link{read_json_conn}()},
\code{\link{read_json_file}()},
\code{\link{read_json_raw}()},
\code{\link{read_json_str}()},
\code{\link{read_ndjson_file}()},
\code{\link{read_ndjson_raw}()},
\code{\link{read_ndjson_str}()}
}
\concept{JSON Parsers}
