% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson.R
\name{read_geojson_str}
\alias{read_geojson_str}
\alias{read_geojson_file}
\title{Load GeoJSON as \code{sf} object}
\usage{
read_geojson_str(str, opts = list(), ...)

read_geojson_file(filename, opts = list(), ...)
}
\arguments{
\item{str}{single character string containing GeoJSON}

\item{opts}{named list of options. Usually created with \code{from_geojson_opts()}.
Default: empty \code{list()} to use the default options.}

\item{...}{any extra named options override those in \code{opts}}

\item{filename}{filename}
}
\value{
\code{sf} object
}
\description{
Load GeoJSON as \code{sf} object
}
