% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson.R
\name{to_geojson_opts}
\alias{to_geojson_opts}
\title{Options for writing from \code{sf} object to \code{GeoJSON}}
\usage{
to_geojson_opts()
}
\value{
named list of options
}
\description{
Currently no options available.
}
