% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson.R
\name{to_geojson_str}
\alias{to_geojson_str}
\alias{to_geojson_file}
\title{Write SF to geojson string}
\usage{
to_geojson_str(x, opts = list(), ...)

to_geojson_file(x, filename, opts = list(), ...)
}
\arguments{
\item{x}{\code{sf} object. Supports \code{sf} or \code{sfc}}

\item{opts}{named list of options. Usually created with \code{to_geojson_opts()}.
Default: empty \code{list()} to use the default options.}

\item{...}{any extra named options override those in \code{opts}}

\item{filename}{filename}
}
\value{
character string containing json
}
\description{
Write SF to geojson string
}
