% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{to_json_str}
\alias{to_json_str}
\title{Convert R object to JSON string}
\usage{
to_json_str(x, opts = list(), ...)
}
\arguments{
\item{x}{the object to be encoded}

\item{opts}{Named list of serialization options. Usually created by \code{\link[=to_opts]{to_opts()}}}

\item{...}{Other named options can be used to override any options in \code{opts}.
The valid named options are identical to arguments to \code{\link[=to_opts]{to_opts()}}}
}
\value{
Character string
}
\description{
Convert R object to JSON string
}
\examples{
\dontrun{
to_json_str(iris, pretty = TRUE)
to_json_str(iris, opts = to_opts(auto_unbox = FALSE))
}

}
\seealso{
Other JSON Serializer: 
\code{\link{to_json_file}()}
}
\concept{JSON Serializer}
