% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson.R
\name{write_geojson_raw}
\alias{write_geojson_raw}
\title{Write SF to GeoJSON raw vector}
\usage{
write_geojson_raw(x, opts = list(), ..., json_opts = list())
}
\arguments{
\item{x}{\code{sf} object. Supports \code{sf} or \code{sfc}}

\item{opts}{named list of options. Usually created with \code{opts_write_geojson()}.
Default: empty \code{list()} to use the default options.}

\item{...}{any extra named options override those in \code{opts}}

\item{json_opts}{Named list of vanilla JSON options as used by \code{write_json_str()}.
This is usually created with \code{opts_write_json()}.  Default value is
an empty \code{list()} which means to use all the default JSON writing
options which is usually the correct thing to do when writing GeoJSON.}
}
\value{
Raw vector containing GeoJSON.
}
\description{
Coordinate reference system is always WGS84 in accordance with GeoJSON RFC.
}
\examples{
geojson_file <- system.file("geojson-example.json", package = 'yyjsonr')
sf <- read_geojson_file(geojson_file)
js <- write_geojson_raw(sf)
}
