% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson.R
\name{write_geojson_str}
\alias{write_geojson_str}
\alias{write_geojson_file}
\title{Write SF to geojson string}
\usage{
write_geojson_str(x, opts = list(), ..., digits = -1)

write_geojson_file(x, filename, opts = list(), ..., digits = -1)
}
\arguments{
\item{x}{\code{sf} object. Supports \code{sf} or \code{sfc}}

\item{opts}{named list of options. Usually created with \code{opts_write_geojson()}.
Default: empty \code{list()} to use the default options.}

\item{...}{any extra named options override those in \code{opts}}

\item{digits}{decimal places to keep for floating point numbers. Default: -1.
Positive values specify number of decimal places. Using zero will
write the numeric value as an integer. Values less than zero mean that
the floating point value should be written as-is (the default).}

\item{filename}{filename}
}
\value{
character string containing json
}
\description{
Write SF to geojson string
}
