% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yyjson_version}
\alias{yyjson_version}
\title{Version number of 'yyjson' C library}
\usage{
yyjson_version()
}
\value{
Version of included yyjson C library as a string
}
\description{
Version number of 'yyjson' C library
}
\examples{
yyjson_version()
}
