% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson.R
\name{opts_write_geojson}
\alias{opts_write_geojson}
\title{Options for writing from \code{sf} object to \code{GeoJSON}}
\usage{
opts_write_geojson()
}
\value{
named list of options
}
\description{
Currently no options available.
}
