% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-ndjson.R
\name{read_dataset_ndjson_str}
\alias{read_dataset_ndjson_str}
\title{Parse an Dataset-NDJSON string to a data.frame}
\usage{
read_dataset_ndjson_str(str, nskip = 1, colspec = NULL, opts = list(), ...)
}
\arguments{
\item{str}{a single character string}

\item{nskip}{Number of records to skip before starting to read. Default: 0
(skip no data)}

\item{colspec}{data.frame of column spec. Must have a row for each data item
expected in the dataset.  Must include a 'name' and 'dataType' column}

\item{opts}{Named list of options for parsing. Usually created by \code{opts_read_json()}}

\item{...}{Other named options can be used to override any options in \code{opts}.
The valid named options are identical to arguments to \code{\link[=opts_read_json]{opts_read_json()}}}
}
\value{
Dataset-NDJSON data read into R as data.frame
}
\description{
This is quite a strict parser and a colspec \emph{must} be given and match
the data stream.
}
\seealso{
Other JSON Parsers: 
\code{\link{read_geojson_conn}()},
\code{\link{read_json_conn}()},
\code{\link{read_json_file}()},
\code{\link{read_json_raw}()},
\code{\link{read_json_str}()},
\code{\link{read_ndjson_file}()},
\code{\link{read_ndjson_raw}()},
\code{\link{read_ndjson_str}()}
}
\concept{JSON Parsers}
