% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{write_json_file}
\alias{write_json_file}
\title{Convert R object to JSON file}
\usage{
write_json_file(x, filename, opts = list(), ...)
}
\arguments{
\item{x}{the object to be encoded}

\item{filename}{filename}

\item{opts}{Named list of serialization options. Usually created by \code{\link[=to_opts]{to_opts()}}}

\item{...}{Other named options can be used to override any options in \code{opts}.
The valid named options are identical to arguments to \code{\link[=to_opts]{to_opts()}}}
}
\description{
Convert R object to JSON file
}
\examples{
\dontrun{
write_json_str(iris, filename = "iris.json")
}

}
\seealso{
Other JSON Serializer: 
\code{\link{write_json_str}()}
}
\concept{JSON Serializer}
