% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndjson.R
\name{write_ndjson_file}
\alias{write_ndjson_file}
\alias{write_ndjson_str}
\title{Write list or data.frame object to ndjson}
\usage{
write_ndjson_file(x, filename, opts = list(), ...)

write_ndjson_str(x, opts = list(), ...)
}
\arguments{
\item{x}{data.frame or list}

\item{filename}{filename}

\item{opts}{Named list of serialization options. Usually created by \code{\link[=opts_write_json]{opts_write_json()}}}

\item{...}{Other named options can be used to override any options in \code{opts}.
The valid named options are identical to arguments to \code{\link[=opts_write_json]{opts_write_json()}}}
}
\description{
Write list or data.frame object to ndjson
}
\examples{
\dontrun{
write_json_str(iris, filename = "iris.json")
}

}
\seealso{
Other JSON Serializer: 
\code{\link{write_json_str}()}
}
